Added more REST implementations.
Added site/REST/search.php, that returns a json encoded string containing
the tablets, scores, and total number of results returned.

Added site/REST/name.php, that returns a json encoded string containing
the names that are present in the search.

Added site/REST/dates.php, that returns a json encoded string containing
the dates that are present in the search.

Update site/REST/tablet.php to verify the provided tablet_id is an
integer.

Added popups chart of dates and names.